스터디 좋아요 파일구조를 개선하라
스터디 좋아요 서비스를 따로 분리합니다. 한곳에 너무 많은 로직을 넣지
않고 분산시키기위해서입니다
