+++ app/src/main/java/com/example/bookclub/controller/api/StudyApiController.java
@@ -1,5 +1,6 @@
 package com.example.bookclub.controller.api;
 
+import com.example.bookclub.application.StudyLikeService;
 import com.example.bookclub.application.StudyService;
 import com.example.bookclub.domain.Account;
 import com.example.bookclub.domain.Study;
@@ -26,9 +27,12 @@ import java.util.List;
 @RequestMapping("/api/study")
 public class StudyApiController {
     private final StudyService studyService;
+    private final StudyLikeService studyLikeService;
 
-    public StudyApiController(StudyService studyService) {
+    public StudyApiController(StudyService studyService,
+                              StudyLikeService studyLikeService) {
         this.studyService = studyService;
+        this.studyLikeService = studyLikeService;
     }
 
     @GetMapping
@@ -80,6 +84,6 @@ public class StudyApiController {
     @PostMapping("/like/{studyId}")
     public Long like(@AuthenticationPrincipal UserAccount userAccount,
                      @PathVariable Long studyId) {
-        return studyService.like(userAccount, studyId);
+        return studyLikeService.like(userAccount, studyId);
     }
 }
